requirements.txt
click==7.1.2
pytest==5.4.2
PyYAML==5.3.1
tabulate~=0.8.7